requirements.txt
flask
joblib
requests
pandas
numpy
scikit-learn
geopy
xgboost
reportlab
gunicorn
python-dotenv
